Version 1.2.1 - Manage Player Patch
Hello @everyone , with 2 pings in one day, I'm just keeping this one short. For the /manage_player command, where a DM awards/sets/or manages XP, there is now logging! The logging will be in the same channel that the level ups are in. Alongside this, there is also a new memo option, if you would like to leave a footnote on why the command was used.