Verson 1.2.3 - QOL Mod Patch
Hello @everyone , this announcement is for the mods that use the XPholder, and players wont see much of it for now. The following commands :
/approve
/manage_player
/register
/manage_server
have all been updated to now support a set of options, instead of having people guess what number I want them to use. I hope that this QOL makes moderating your XPholder more pleasant!