Version 2.3 - XP Share
     Hello @everyone , thank you to all the servers who are using XPholder. With some popular request, there is now a new command called /set_xp_roles. The 'xp freeze' as well as the new 'xp share' role options are toggle-able through this command! In addition, the /awared_xp command now has an undo button, incase of accidental XP awards.

Commands [ Everyone - New]
/toggle_xp_roles - allows for the user to add or remove : XP freeze, XP Share

Commands [ Everyone - Update]
/request_xp - bug when requesting CXP, command would fail
/xp - XP freeze button moved to /toggle_xp_roles

Commands [ Mod - Update]
/award_xp - now supports undo
/view_game_rules - now supports xp share

Commands [ Owner - Update]
/register - now adds xp share
/edit_config - now supports xp share

IMPORT PLEASE ALL SERVER OWNERS READ
To be respectful to servers, I did not have XPholder auto generate the new 'xp share' role. This is because I want to make sure that if a server is adding this new feature that it is a conscious decision. This is a non essential feature, and those servers who wish to use it can add the role with the /edit_config command

Once again, thank you all for you love and support of the bot, if you have friends that are interested, feel free to invite them to the dev server with this link :  https://discord.gg/gawCJ6EYcb